evm_mine: Add option to not emit `new block header` event
By default, upon mining a new block, an event is emitted to inform the
outside world about that fact. However, for simulating chain
reorganizations, it is useful to suppress the `new block` event for all
but the select blocks, typically the latest block in the new chain.

This commit addresses this concern by introducing the following changes:

1. In `BlockchainDouble.putBlock`, the `new block` event will be emitted
only if `emitNewHeaderEvent == true`.

2. `BlockchainDouble.putBlock` and `BlockchainDouble.processBlock` now
accept and additional parameter `emitNewHeaderEvent`, which defaults to
`true`, to be consistent with the existing behavior.

3. `evm_mine`, `BlockchainDouble.processNextBlock`, and
`StateManager.processBlock` now take an object as a parameter. This
allows those functions to easily accept any number of optional
parameters. Otherwise, it would be quite difficult to manage several
optional parameters that are not the rightmost parameters in a function
signature.

4. `timestamp` and `emitNewHeaderEvent` are treated as optional
parameters with default values of `null` and `true`, respectively. This
results in exactly the same behavior that existed prior to this commit.
Changes introduced by this commit alter the existing behavior only if
`emitNewHeaderEvent == false` is provided.

5. Wherever these functions are called in the source code, their
interfaces have been adjusted accordingly, preserving the previously
implicit defaults.